add third recipe, finalize README and edits
These changes add the thrid and final recipe to the project which is a
recipe for friend chicken. This recipe required an image so that is
included. Changed index to link to the recipe as well changed the title
of the homepage. Completed README message with a description of the full
project details. Edited typo of waffle images alt text. Changed the
description of the garlic bread recipe to make it less possessive.